Refactor: Rename SignDocument screen and related classes to DocumentSign
This commit renames the `SignDocument` screen and its associated classes to `DocumentSign` for better clarity and consistency.

The following changes were made:
- Renamed `RouterContract.SignDocument` to `RouterContract.DocumentSign`.
- Renamed files and classes within `dashboard-feature` from `sign` to `document_sign`.
- Updated references to the screen and classes in `HomeViewModel`, `DocumentSignInteractor`, and `Graph.kt`.
- Updated string resource ID from `signDocumentButtonId` to `documentSignButtonId`.
