+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/interactor/DocumentSignInteractor.kt
@@ -18,7 +18,7 @@ package eu.europa.ec.dashboardfeature.interactor
 
 import android.content.Context
 import android.net.Uri
-import eu.europa.ec.dashboardfeature.ui.sign.model.SignDocumentButtonUi
+import eu.europa.ec.dashboardfeature.ui.document_sign.model.DocumentSignButtonUi
 import eu.europa.ec.eudi.rqesui.infrastructure.DocumentUri
 import eu.europa.ec.eudi.rqesui.infrastructure.EudiRQESUi
 import eu.europa.ec.resourceslogic.R
@@ -30,7 +30,7 @@ import eu.europa.ec.uilogic.component.ListItemTrailingContentDataUi
 
 interface DocumentSignInteractor {
     fun launchRqesSdk(context: Context, uri: Uri)
-    fun getItemUi(): SignDocumentButtonUi
+    fun getItemUi(): DocumentSignButtonUi
 }
 
 class DocumentSignInteractorImpl(
@@ -44,8 +44,8 @@ class DocumentSignInteractorImpl(
         )
     }
 
-    override fun getItemUi(): SignDocumentButtonUi {
-        return SignDocumentButtonUi(
+    override fun getItemUi(): DocumentSignButtonUi {
+        return DocumentSignButtonUi(
             data = ListItemDataUi(
                 itemId = resourceProvider.getString(R.string.document_sign_select_document_button_id),
                 mainContentData = ListItemMainContentDataUi.Text(
