+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/router/Graph.kt
@@ -25,9 +25,9 @@ import androidx.navigation.navArgument
 import androidx.navigation.navDeepLink
 import eu.europa.ec.dashboardfeature.BuildConfig
 import eu.europa.ec.dashboardfeature.ui.dashboard.DashboardScreen
+import eu.europa.ec.dashboardfeature.ui.document_sign.DocumentSignScreen
 import eu.europa.ec.dashboardfeature.ui.documents.detail.DocumentDetailsScreen
 import eu.europa.ec.dashboardfeature.ui.settings.SettingsScreen
-import eu.europa.ec.dashboardfeature.ui.sign.DocumentSignScreen
 import eu.europa.ec.dashboardfeature.ui.transactions.detail.TransactionDetailsScreen
 import eu.europa.ec.uilogic.navigation.DashboardScreens
 import eu.europa.ec.uilogic.navigation.ModuleRoute
@@ -73,7 +73,7 @@ fun NavGraphBuilder.featureDashboardGraph(navController: NavController) {
         }
 
         composable(
-            route = DashboardScreens.SignDocument.screenRoute
+            route = DashboardScreens.DocumentSign.screenRoute
         ) {
             DocumentSignScreen(navController, koinViewModel())
         }
